ai-image-classifier/backend/requirements.txt
tensorflow==2.16.2
pillow==10.0.0
numpy==1.24.3
flask==2.3.2
matplotlib==3.7.1
scikit-learn==1.2.2